backend/requirements.txt
asgiref==3.8.1
astroid==3.3.8
dill==0.3.9
Django==5.1.6
djangorestframework==3.15.2
isort==6.0.0
mccabe==0.7.0
platformdirs==4.3.6
pylint==3.3.4
sqlparse==0.5.3
tomlkit==0.13.2
psycopg
djangorestframework-simplejwt